ui updates, posts app created, Post model migrated
