+++ 5_Django_study_project/core/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class TestappConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "core"
