+++ 5_Django_study_project/posts/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from .models import Post
+
+# Register your models here.
+
+admin.site.register(Post)
+
