+++ 5_Django_study_project/posts/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class PostsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "posts"
