+++ 5_Django_study_project/posts/models.py
@@ -1,0 +1,11 @@
+from django.db import models
+
+# Create your models here.
+
+
+class Post(models.Model):
+    title = models.CharField(max_length=255)
+    body = models.TextField()
+
+    def __str__(self):
+        return self.title[:50]
