+++ 5_Django_study_project/static/css/base.css
@@ -16,7 +16,8 @@ body {
 html,
 body {
   height: 100%;
-	overflow: hidden;
+    overflow: hidden;
+    background-color: #396060;
 }
 
 canvas {
@@ -33,6 +34,7 @@ canvas {
 
 .content-wrap {
     width: calc(100% - 200px);
+    max-width: 1400px;
   margin: 0 auto;
 }
 
@@ -42,4 +44,52 @@ canvas {
     text-align: center;
     left: 50%;
     margin-top: 20px;
+}
+
+/*=============== Shapes ===============*/
+
+.hero__shape-small,
+.hero__shape-big-1,
+.hero__shape-big-2{
+    position: relative;
+    border-radius: 50%;
+    filter: blur(140px);
+}
+
+.hero__shape-small{
+    width: 350px;
+    height: 350px;
+    background-color: rgba(106, 20, 255, 0.5);
+    top: -20rem;
+    left: -5rem;
+}
+
+.hero__shape-big-1,
+.hero__shape-big-2{
+    width: 300px;
+    height: 300px;
+}
+
+.hero__shape-big-1{
+    background-color: rgba(20, 255, 247, 0.60);
+    top: -25rem;
+    left: 75rem;
+}
+
+.hero__shape-big-2{
+    background-color: rgba(255, 118, 20, 0.58);
+    top: 20rem;
+    left: 27rem;
+}
+
+.hero__shape-bg{
+    position: absolute;
+    z-index: 0;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100vh;
+    object-fit: cover;
+    mix-blend-mode: soft-light;
+    opacity: .15;
 }
