+++ 5_Django_study_project/static/css/footer.css
@@ -1,7 +1,11 @@
 
 
 footer {
-  background-color: burlywood;
+  position: fixed;
+    bottom: 0;
+  left: 0;
+    width: 100%;
+
   text-align: center;
   height: 100px;
 }
