+++ 5_Django_study_project/static/js/spline.js
@@ -2,4 +2,4 @@ import { Application } from '@splinetool/runtime';
 
 const canvas = document.getElementById('canvas3d');
 const app = new Application(canvas);
-app.load('https://prod.spline.design/M5gV0GWpS0eLRamU/scene.splinecode');
+app.load('https://prod.spline.design/MB1gX6RMZAr4WZ4M/scene.splinecode');
