+++ 5_Django_study_project/templates/testapp/base.html
@@ -24,7 +24,9 @@
 
                     <div class="menu-container">
                         <div class="logo-container">
-                            <img src="{% static "img/logo.png" %}" alt="logo">
+{#                            <a href="{% url 'home' %}"><img src="{% static "img/logo.png" %}" alt="logo"></a>#}
+                            <a href="{% url 'home' %}"><canvas id="canvas3d"></canvas></a>
+
                         </div>
                         <ul class="menu-left">
                             <li><a href="{% url 'home' %}">Home</a></li>
@@ -46,9 +48,9 @@
                                         Welcome,&nbsp; <span style="color: mediumvioletred">{{ user.username }}</span>
                                     </div>
                                 </li>
-                                <li><a href="{% url 'logout' %}">logout</a></li>
+                                <li><a href="{% url 'logout' %}">Logout</a></li>
                             {% else %}
-                                <li><a href="{% url 'login' %}">login</a></li>
+                                <li><a href="{% url 'login' %}">Login</a></li>
                                 <li><a href="{% url 'register' %}">Register</a></li>
                             {% endif %}
                         </ul>
@@ -64,43 +66,48 @@
             <h1>Test</h1>
         {% endblock content%}
 
+{% block hero %}
+{% endblock %}
 
+{% block paragraph %}
+        This is the BASE paragraph block
+    {% block inner_paragraph %}
+        <br>
+        This is the BASE paragraph INNER block
+    {% endblock inner_paragraph %}
+    <br>Continue BASE paragraph block
+{% endblock paragraph %}
 
-        {% block paragraph %}
-                This is the BASE paragraph block
-            {% block inner_paragraph %}
-                <br>
-                This is the BASE paragraph INNER block
-            {% endblock inner_paragraph %}
-            <br>Continue BASE paragraph block
-        {% endblock paragraph %}
+{% block posts %}
+{% endblock %}
 
+{#{% block canvas %}#}
+{#<canvas id="canvas3d"></canvas>#}
+{#{% endblock canvas %}#}
 
 
   </div>
 
-{#<canvas id="canvas3d">#}
-{#</canvas>#}
 
 
-<!-- Import maps polyfill -->
-<!-- Remove this when import maps will be widely supported -->
-<script async src="https://unpkg.com/es-module-shims@1.3.6/dist/es-module-shims.js"></script>
-<script type="importmap">
-  {
-    "imports": {
-      "@splinetool/runtime": "https://unpkg.com/@splinetool/runtime@0.9.480/build/runtime.js"
-    }
-  }
-</script>
+
+    <!-- Import maps polyfill -->
+    <!-- Remove this when import maps will be widely supported -->
+    <script async src="https://unpkg.com/es-module-shims@1.3.6/dist/es-module-shims.js"></script>
+    <script type="importmap">
+      {
+        "imports": {
+          "@splinetool/runtime": "https://unpkg.com/@splinetool/runtime@0.9.483/build/runtime.js"
+        }
+      }
+    </script>
 <script type="module" src="{% static 'js/spline.js' %}"></script>
+
 
 
 {% block footer %}
     {% render_footer 'AwesomeCo' %}
 {% endblock footer%}
-
-
 
 </div>
