--- 5_Django_study_project/testapp/admin.py
@@ -1,3 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
