--- 5_Django_study_project/testapp/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class TestappConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "testapp"
